Remove Edit option from the service catalog
