+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/site/public/locales/en.json
@@ -1092,39 +1092,6 @@
   "ServiceCatalog.Listing.Delete.confirm": "Confirm Delete",
   "ServiceCatalog.Listing.Delete.ok.confirm": "Are you sure you want to delete the service {service} ?",
   "ServiceCatalog.Listing.Delete.ok.yes": "Yes",
-  "ServiceCatalog.Listing.Edit.service.display.name.empty": "Service display name is empty",
-  "ServiceCatalog.Listing.Edit.service.display.name.invalid.character": "Service display name contains one or more illegal characters",
-  "ServiceCatalog.Listing.Edit.service.display.name.too.long": "Service display name is too long",
-  "ServiceCatalog.Listing.Edit.service.url.empty": "Service Url is empty",
-  "ServiceCatalog.Listing.EditableOverview.back.btn": "Go Back",
-  "ServiceCatalog.Listing.EditableOverview.cancel.btn": "Cancel",
-  "ServiceCatalog.Listing.EditableOverview.created.time": "Created Time",
-  "ServiceCatalog.Listing.EditableOverview.definition.type": "Schema Type",
-  "ServiceCatalog.Listing.EditableOverview.description.ht": "Description of the Service",
-  "ServiceCatalog.Listing.EditableOverview.display.name": "{serviceDisplayName}",
-  "ServiceCatalog.Listing.EditableOverview.dn": "Display name of the service",
-  "ServiceCatalog.Listing.EditableOverview.edit.breadcrumb": "Edit",
-  "ServiceCatalog.Listing.EditableOverview.edit.btn": "Edit",
-  "ServiceCatalog.Listing.EditableOverview.error.loading.service": "Error while loading service",
-  "ServiceCatalog.Listing.EditableOverview.error.retrieve.settings.data": "Error while retrieving settings data",
-  "ServiceCatalog.Listing.EditableOverview.error.update": "Error while updating service",
-  "ServiceCatalog.Listing.EditableOverview.mutual.ssl": "Mutual SSL",
-  "ServiceCatalog.Listing.EditableOverview.mutual.ssl.disabled": "Disabled",
-  "ServiceCatalog.Listing.EditableOverview.mutual.ssl.enable.label": "Enable",
-  "ServiceCatalog.Listing.EditableOverview.mutual.ssl.enabled": "Enabled",
-  "ServiceCatalog.Listing.EditableOverview.no.description.text": "Description of the Service",
-  "ServiceCatalog.Listing.EditableOverview.parent.breadcrumb": "Service Catalog",
-  "ServiceCatalog.Listing.EditableOverview.readonly.breadcrumb": "Overview",
-  "ServiceCatalog.Listing.EditableOverview.save.btn": "Save",
-  "ServiceCatalog.Listing.EditableOverview.security.type": "Security Type",
-  "ServiceCatalog.Listing.EditableOverview.service.description": "{description}",
-  "ServiceCatalog.Listing.EditableOverview.service.successful.update": "Service updated successfully!",
-  "ServiceCatalog.Listing.EditableOverview.service.type.async.tooltip": "Async API Service",
-  "ServiceCatalog.Listing.EditableOverview.service.type.graphql.tooltip": "GraphQL Service",
-  "ServiceCatalog.Listing.EditableOverview.service.type.rest.tooltip": "REST Service",
-  "ServiceCatalog.Listing.EditableOverview.service.type.soap.tooltip": "SOAP Service",
-  "ServiceCatalog.Listing.EditableOverview.service.url": "Service URL",
-  "ServiceCatalog.Listing.EditableOverview.service.version": "{serviceVersion}",
   "ServiceCatalog.Listing.Listing.created.time": "Created Time",
   "ServiceCatalog.Listing.Listing.error.delete": "Error while deleting service",
   "ServiceCatalog.Listing.Listing.error.loading": "Error while loading services",
@@ -1142,8 +1109,24 @@
   "ServiceCatalog.Listing.Onboarding.get.started": "Get Started",
   "ServiceCatalog.Listing.Onboarding.welcome.description": "The Service Catalog enables API-First Integration. Through this, integration services are made discoverable to the API Management layer so that API proxies can directly be created using them.",
   "ServiceCatalog.Listing.Onboarding.welcome.msg": "Welcome to WSO2 API Manager",
-  "ServiceCatalog.Listing.Overview.service.definition.type.empty": "Schema Type is empty",
-  "ServiceCatalog.Listing.Overview.service.security.type.empty": "Security Type is empty",
+  "ServiceCatalog.Listing.Overview.back.btn": "Go Back",
+  "ServiceCatalog.Listing.Overview.created.time": "Created Time",
+  "ServiceCatalog.Listing.Overview.definition.type": "Schema Type",
+  "ServiceCatalog.Listing.Overview.display.name": "{serviceDisplayName}",
+  "ServiceCatalog.Listing.Overview.error.loading.service": "Error while loading service",
+  "ServiceCatalog.Listing.Overview.mutual.ssl": "Mutual SSL",
+  "ServiceCatalog.Listing.Overview.mutual.ssl.disabled": "Disabled",
+  "ServiceCatalog.Listing.Overview.mutual.ssl.enabled": "Enabled",
+  "ServiceCatalog.Listing.Overview.parent.breadcrumb": "Service Catalog",
+  "ServiceCatalog.Listing.Overview.readonly.breadcrumb": "Overview",
+  "ServiceCatalog.Listing.Overview.security.type": "Security Type",
+  "ServiceCatalog.Listing.Overview.service.description": "{description}",
+  "ServiceCatalog.Listing.Overview.service.type.async.tooltip": "Async API Service",
+  "ServiceCatalog.Listing.Overview.service.type.graphql.tooltip": "GraphQL Service",
+  "ServiceCatalog.Listing.Overview.service.type.rest.tooltip": "REST Service",
+  "ServiceCatalog.Listing.Overview.service.type.soap.tooltip": "SOAP Service",
+  "ServiceCatalog.Listing.Overview.service.url": "Service URL",
+  "ServiceCatalog.Listing.Overview.service.version": "{serviceVersion}",
   "ServiceCatalog.Listing.Usages.api.context": "Context",
   "ServiceCatalog.Listing.Usages.api.name": "API Name",
   "ServiceCatalog.Listing.Usages.api.version": "Version",
