+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/site/public/locales/fr.json
@@ -1092,39 +1092,6 @@
   "ServiceCatalog.Listing.Delete.confirm": "",
   "ServiceCatalog.Listing.Delete.ok.confirm": "",
   "ServiceCatalog.Listing.Delete.ok.yes": "",
-  "ServiceCatalog.Listing.Edit.service.display.name.empty": "",
-  "ServiceCatalog.Listing.Edit.service.display.name.invalid.character": "",
-  "ServiceCatalog.Listing.Edit.service.display.name.too.long": "",
-  "ServiceCatalog.Listing.Edit.service.url.empty": "",
-  "ServiceCatalog.Listing.EditableOverview.back.btn": "",
-  "ServiceCatalog.Listing.EditableOverview.cancel.btn": "",
-  "ServiceCatalog.Listing.EditableOverview.created.time": "",
-  "ServiceCatalog.Listing.EditableOverview.definition.type": "",
-  "ServiceCatalog.Listing.EditableOverview.description.ht": "",
-  "ServiceCatalog.Listing.EditableOverview.display.name": "",
-  "ServiceCatalog.Listing.EditableOverview.dn": "",
-  "ServiceCatalog.Listing.EditableOverview.edit.breadcrumb": "",
-  "ServiceCatalog.Listing.EditableOverview.edit.btn": "",
-  "ServiceCatalog.Listing.EditableOverview.error.loading.service": "",
-  "ServiceCatalog.Listing.EditableOverview.error.retrieve.settings.data": "",
-  "ServiceCatalog.Listing.EditableOverview.error.update": "",
-  "ServiceCatalog.Listing.EditableOverview.mutual.ssl": "",
-  "ServiceCatalog.Listing.EditableOverview.mutual.ssl.disabled": "",
-  "ServiceCatalog.Listing.EditableOverview.mutual.ssl.enable.label": "",
-  "ServiceCatalog.Listing.EditableOverview.mutual.ssl.enabled": "",
-  "ServiceCatalog.Listing.EditableOverview.no.description.text": "",
-  "ServiceCatalog.Listing.EditableOverview.parent.breadcrumb": "",
-  "ServiceCatalog.Listing.EditableOverview.readonly.breadcrumb": "",
-  "ServiceCatalog.Listing.EditableOverview.save.btn": "",
-  "ServiceCatalog.Listing.EditableOverview.security.type": "",
-  "ServiceCatalog.Listing.EditableOverview.service.description": "",
-  "ServiceCatalog.Listing.EditableOverview.service.successful.update": "",
-  "ServiceCatalog.Listing.EditableOverview.service.type.async.tooltip": "",
-  "ServiceCatalog.Listing.EditableOverview.service.type.graphql.tooltip": "",
-  "ServiceCatalog.Listing.EditableOverview.service.type.rest.tooltip": "",
-  "ServiceCatalog.Listing.EditableOverview.service.type.soap.tooltip": "",
-  "ServiceCatalog.Listing.EditableOverview.service.url": "",
-  "ServiceCatalog.Listing.EditableOverview.service.version": "",
   "ServiceCatalog.Listing.Listing.created.time": "",
   "ServiceCatalog.Listing.Listing.error.delete": "",
   "ServiceCatalog.Listing.Listing.error.loading": "",
@@ -1142,8 +1109,24 @@
   "ServiceCatalog.Listing.Onboarding.get.started": "",
   "ServiceCatalog.Listing.Onboarding.welcome.description": "",
   "ServiceCatalog.Listing.Onboarding.welcome.msg": "",
-  "ServiceCatalog.Listing.Overview.service.definition.type.empty": "",
-  "ServiceCatalog.Listing.Overview.service.security.type.empty": "",
+  "ServiceCatalog.Listing.Overview.back.btn": "",
+  "ServiceCatalog.Listing.Overview.created.time": "",
+  "ServiceCatalog.Listing.Overview.definition.type": "",
+  "ServiceCatalog.Listing.Overview.display.name": "",
+  "ServiceCatalog.Listing.Overview.error.loading.service": "",
+  "ServiceCatalog.Listing.Overview.mutual.ssl": "",
+  "ServiceCatalog.Listing.Overview.mutual.ssl.disabled": "",
+  "ServiceCatalog.Listing.Overview.mutual.ssl.enabled": "",
+  "ServiceCatalog.Listing.Overview.parent.breadcrumb": "",
+  "ServiceCatalog.Listing.Overview.readonly.breadcrumb": "",
+  "ServiceCatalog.Listing.Overview.security.type": "",
+  "ServiceCatalog.Listing.Overview.service.description": "",
+  "ServiceCatalog.Listing.Overview.service.type.async.tooltip": "",
+  "ServiceCatalog.Listing.Overview.service.type.graphql.tooltip": "",
+  "ServiceCatalog.Listing.Overview.service.type.rest.tooltip": "",
+  "ServiceCatalog.Listing.Overview.service.type.soap.tooltip": "",
+  "ServiceCatalog.Listing.Overview.service.url": "",
+  "ServiceCatalog.Listing.Overview.service.version": "",
   "ServiceCatalog.Listing.Usages.api.context": "",
   "ServiceCatalog.Listing.Usages.api.name": "",
   "ServiceCatalog.Listing.Usages.api.version": "",
